achievement icon update

+++ scripts/general/global.gd
@@ -684,7 +684,7 @@ func return_to_menu(delete = false, display_stats = false) -> void:
 	else:
 		call_deferred("restart_game")
 	
-	if delete:
+	if delete and current_room_node.TYPE != "tutorial":
 		ChumsManager.add_bonus_chum(Global.game_save_id) #Gives you a random chum (not boss/prisoner) for your next run.
 		SaverLoader.delete_save(game_save_id)
